lab6/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
python-multipart==0.0.6